+++ src/app/app-routing.module.ts
@@ -3,9 +3,9 @@ import { Routes, RouterModule } from '@angular/router';
 import {GoodDetailComponent} from './components/good-detail/good-detail.component';
 import { HomeComponent } from './home/home.component';
 import { GoodListComponent } from './components/good-list/good-list.component';
-import { FormContractComponent } from './components/form-contract/form-contract.component';
+import {FormContractComponent} from './form/form-contract/form-contract.component'
 import {SellFormComponent} from './form/sell-form/sell-form.component';
-import { ContractDetailsComponent } from './contract-details/contract-details.component';
+
 
 const routes: Routes = [
   { path: '', redirectTo: '/home', pathMatch: 'full'},
@@ -13,8 +13,7 @@ const routes: Routes = [
   { path: 'good_details/:id', component: GoodDetailComponent },
   { path: 'goods', component: GoodListComponent },
   { path: 'contractform', component: FormContractComponent},
-  { path: 'sell', component: SellFormComponent},
-  { path: 'contractdetails/:id', component : ContractDetailsComponent}
+  { path: 'sell', component: SellFormComponent}
 
 ];
