--- src/app/contract-details/contract-details.component.html
@@ -1,20 +1,0 @@
-<h1>Contract of agreement<h1>
-  <div>
-    Congratulations! The application for "{{contract.good.nameOfGood}}" referenced n°{{contract.good.id}} was accepted. Hence,
-     {{contract.estateAgent}}, on behalf of AmazingEstateAgency would like to offer you a contract.
-  </div>
-  <div>
-    <h2>Good details</h2>
-    Mr/Mrs {{contract.client.name}} is interested in <span *ngIf="contract.typeOfContract === 'TORENT' "; else buy>renting</span> <span #buy>buying</span> the good
-    referenced {{contract.good.id}}.
-    Size: {{contract.good.area}}
-    Address: {{contract.good.address}}
-    <span *ngIf="contract.typeOfContract === 'TORENT' "; else buy>
-      Montly rate : {{contract.good.price}}
-    </span>
-    <span #buy>
-      Price : {{contract.good.price}}
-    </span>
-  </div>
-
-
